+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 See the [keepachangelog.com description](https://keepachangelog.com/en/1.0.0/).
 
+## 0.2.0 - 2023-08-02
+
+* Added
+  * Yamaha files scanner window
+  * .Tou files reader
+* Changed
+* Deprecated
+* Removed
+* Fixed
+
 ## 0.1.0 - 2023-07-31
 
 * Added